Move SpanData to export package. Move implementation of export package class to export package.
