+++ api/src/main/java/io/opencensus/trace/export/SpanExporter.java
@@ -14,7 +14,6 @@
 package io.opencensus.trace.export;
 
 import io.opencensus.trace.Span;
-import io.opencensus.trace.SpanData;
 import io.opencensus.trace.base.TraceOptions;
 import java.util.Collection;
 import java.util.logging.Level;
